cast strlen to ssize_t

+++ libxo/libxo.c
@@ -2911,7 +2911,7 @@ xo_format_string_direct (xo_handle_t *xop, xo_buffer_t *xbp,
      */
     if (cp && len > 0 && xo_is_style_text_utf8(xop) && need_enc == have_enc) {
 	const char *np, *ep;
-	ssize_t clen = len < 0 ? strlen(cp) : len;
+	ssize_t clen = len < 0 ? (ssize_t) strlen(cp) : len;
 	for (np = cp, ep = cp + clen; np < ep; np++)
 	    if (xo_is_byte_utf8(*np) || *np == '\\' || *np == '%'
 		|| *np == '{' || *np == '}')
